backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
transformers==4.36.2
torch==2.1.0
pillow==10.2.0
pydantic==1.10.13
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
python-docx==1.1.0
